refactor: extract buildAllClassesConfigMap from Chain
